+++ Changelog.md
@@ -43,6 +43,7 @@
 - Fix: Changing the owner of a Laser Block does not change the owner of connected laser blocks
 - Fix: Adding a module to a Laser Block by rightclicking the block does not add the module to connected laser blocks
 - Fix: Module dupe involving doors
+- Fix: It's possible to connect one's own Password-protected Chest to that of another owner
 - Removed: Some unnecessary menu types. This may result in a "missing registry entries" message showing up, which can be accepted
 - Removed: "allowCodebreakerItem" config option. Disabling the Codebreaker is now achieveable by setting "codebreaker_chance" to a negative value
 - Misc.: More texture and model changes